requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
pydantic>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
